chat-system/backend/requirements.txt
MySQL-python==1.2.3
ipython
channels==1.1.6
asgi_redis
django-cors-headers
